// web/src/generated/src/models/ApiPiConditionOperator.ts
/**
 * System: Condition Operator
 */
export enum ApiPiConditionOperator {
  /** Equal */
  Equal = "Eq",
  /** Greater Than */
  GreaterThan = "Gt",
  /** Greater Than Or Equal */
  GreaterThanOrEqual = "Gte",
  /** In */
  In = "In",
  /** Less Than */
  LessThan = "Lt",
  /** Less Than Or Equal */
  LessThanOrEqual = "Lte",
  /** Not Equal */
  NotEqual = "Ne",
  /** Not In */
  NotIn = "Nin"
}